+++ lib/src/exceptions/branta_payment_exception.dart
@@ -1,6 +1,11 @@
+enum BrantaPaymentExceptionReason {
+  tampered,
+}
+
 class BrantaPaymentException implements Exception {
   final String message;
-  const BrantaPaymentException(this.message);
+  final BrantaPaymentExceptionReason? reason;
+  const BrantaPaymentException(this.message, [this.reason]);
 
   @override
   String toString() => 'BrantaPaymentException: $message';